+++ Sources/TokenStats/UsageHistory/Charts/UsageHistoryCostChart.swift
@@ -169,6 +169,8 @@ struct UsageHistoryCostChart: View {
             Color(red: 0.10, green: 0.65, blue: 0.45)
         case .vertexai:
             Color(red: 0.25, green: 0.52, blue: 0.96)
+        case .amp:
+            Color(red: 0.95, green: 0.31, blue: 0.25)
         }
     }